fix Card/Story Scrolling
This fixes the scrolling of the Card component that contains a list of
stories.

Setting max-height to a specific number will cause a breakpoint on small
screens and cause scrolling to be hidden.

Also, setting `overflow: hidden` on parents elements can cause an effect
of no-scroll, even when `overflow: scroll` is specified on a childe
element.