Updating shapefiles and style.json
Final updates to style.json and shapefiles. However, having a git issue. For some reason Git Desktop is not recognizing lowercase/uppercase changes I've made to some of the files which is resulting in some of the shapefiles not being built. I couldn't figure out how to get that to work. So I added a dir "shapefiles - new" and all of these should be replacing the content in "shapefiles" entirely. Sorry about this :/