make the seeds safe to rerun
Use find_or_create_by() to try to locate an existing seed record or
create it if not found.

For the user record, use where().first_or_create() because we cannot
lookup a user by password, so find them by email and, if not found,
create them and set their password and role.

Signed-off-by: Robb Kidd <robb@thekidds.org>