fixes admin user creation in the seeds
THE COMMAS! The commas turned the three statements into a single
statement that returned an array:

    ["password", "password", 1]

So nothing got set on the attributes for the admin object and so the
user was invalid because the password was too short.

This change removes the commas and updates the code to use
`find_or_create_by!` as documented in the Rails guide.[1] The
band-method version was chosen to that seeding the database would fail
if it can not find the admin user or create a new valid one. Eventually
this admin seed will need to change to not expect all installations to
use the same email for the admin user.

[1] https://guides.rubyonrails.org/v5.2/active_record_querying.html#find-or-build-a-new-object

Signed-off-by: Robb Kidd <robb@thekidds.org>