set nginx container to not start until rails is listening
Change the nginx service in the docker-compose configuration to build
the nginx image instead of using the one directly from Docker Hub. Added
dockerize[1] to our nginx Dockerfile so that it will handle rendering the
template of the nginx configuration *and* wait to start the nginx service
until something is listening on the rails host at tcp:3000.

Also, the nginx service does not need to know the NODE_ENV and RAILS_ENV
values, so they were removed.

Updated the logs script to include logs from nginx so a person might see
that nginx is waiting for rails to be available.

Update logs and start scripts to use the long form of command switch
options so that they are less arcane to readers.

[1] https://github.com/jwilder/dockerize

Signed-off-by: Robb Kidd <robb@thekidds.org>