display an informative error when nginx 502s
502 most likely means the Rails process is not responding
(down or still starting up). Display a meaningful message instead of the
enigmatic "Bad Gateway" no soup for you message.

Changed the nginx service to be built from an image so that the error
page could be added directly to the image. Also added dockerize to handle
the template rendering for the nginx config.

Removed NODE_ENV and RAILS_ENV environment variables as unnecessary for
the nginx service.

Signed-off-by: Robb Kidd <robb@thekidds.org>