Increase timeout to 10 minutes
The default timeout is 60 seconds, which would affect some of the CSV
uploads. I tried restricting the timeout to the `import_csv` routes only
but nginx rule inheritance is a little complicated, so I'm taking the
easy way out by increasing the timeout globally.

While this probably won't be a problem for the NUC local wifi
deployments, this timeout should be restricted to `import_csv` routes if
deployed to the cloud, because it would be easy to DOS the server with a
couple long running requests.