Pause all videos when playing new video
This adds a global JS function to `window` to pause all playing videos
when playing a new video.

The flow goes something like:

- Binds to the `play` event when creating the video component
- When `play` event is triggered, the target of the event is passed to
  `pauseAllVideos()`
- This looks up all `<video>` tags and pauses them (unless it was one
  from the event)

This is admittedly a janky hack that is not very reactive. After
discussing with @mirandawang decided this was Good Enough™ for now, and
if we want to refactor it to better fit into the framework we can.

This decision was a side effect of not me not understanding how to React
properly (suggestions are welcome!).

Fixes #153