Change translate name for the 'place_type'
The task was to translate the 'place_type' and 'region' string from 
marker popup from the map. It turns out that the texts was hard coded. I 
found the key named 'select_place_type' in those files that was exactly 
the translate I needed. So to not create another key for that I just 
renamed the existed one. Also for the record that existed key was only 
used for the story filter in the index page, so it did not interfered.