Add stories sorting in the story list
Add sorting dropdown to the story list. The dropdown contains the
following options - most recent, alphabetical and reveresed
alphabetical.

Change the JSON builder to fetch the created_at attribute for stories.

Fix the line endings in the en.yml file to UNIX, so it is consistent
with the other files.