Update landing for authed user without community
A user without a community should have an opportunity to search for
public facing maps and stories of communities within Terrastories. This
is not yet developed or a feature we've put much thought into, so this
is serving as a placeholder that a search feature is coming.

Note: It may be worth setting this to a default community (like our
terrastories demo one) to keep things somewhat regular in the future.