Terrastories Multi-Instance Support (#572)
* Add Community table

In preparation for Terrastories multi-instance, this adds the new table
to represent a Community in Terrastories.

Currently, this table is standalone and has no relationships to other
models in the application, though that will change as features are built
out.

* Add Community to Users

A given user can belong to a single community.

* Add Community to Places && update seeds/factories

* Add community_id to Themes

* Update landing for authed user without community

A user without a community should have an opportunity to search for
public facing maps and stories of communities within Terrastories. This
is not yet developed or a feature we've put much thought into, so this
is serving as a placeholder that a search feature is coming.

Note: It may be worth setting this to a default community (like our
terrastories demo one) to keep things somewhat regular in the future.

* fixup add Community to User factory

* Update Place.import_csv to belong to Community

With multi-instance, a Place now belongs to a specific community. Rather
than requiring all of the CSVs to provide a Community ID, this updates
the Place.import_csv code to accept a `community` parameter that will
get passed from the controller based on who is logged in.

* Add Community to Speaker and Story

Also update import logic, similar to how Place was done. Since it was
very similar, I did them all at once rather than one at a time.

Additionally: Speaker had a column named "community" — to avoid
collisions, I've renamed this column "speaker_community" just in case we
still use it / want it.

* Fix issues with rebase re: theme

* Update story policy and scope to restrict to stories only in a user's community

* revamp roles

* First pass at new permissions for admin panel

* Allow any auth'd user to access their commmunity page.

* Add second community and fix scoping

* Fix tests

* Revamps style sheets and admin panel look

Warning: only partially functional!

This commit is largely so the changes to the styles persist as I
experiement more with the files and look and feel. This is current
functionally matching to the terrastories website for the header

* Update header so links are full height

* Update translations for Community/Communities

* Update navigation to submenu community theme/settings and users

* Rescue from Pundit::NotAuthorized with flash message

* remove password from show page, and update navigation to go to edit user profile rather than show

* Revamp speakers list view

* remove birthdate and add speaker community to speaker list view

* Revamp Speaker Show profile

* Fix policies: include create/update actions

* Add badge css

* Revamp stories show view

* Add administrate label helpers

* fix list/index policies on adminstrate

* Show Terrastories logo if logo doesn't exist on Community

* Adjust theme policies and items in the view

* Clean up show page

* Rebase issues from update with master

* Fix navigation on mobile/smaller screens

* Fix import button loading icon

* Fix dropdown on mobile nav

* Add missing administrate translations

* Wrap table for mobile/small screens

* Add Annotate gem

* Run annotate on models

* Add Super Admin to users; updates UI for super admin

* Add publich user; always push super admins into admin view without map view

* Fix build theme initialize on community creation

* Update Stories index/dashboard and fixup some of the story styles

* Improve places admin views

* Improve User interface

* Ensure users added to a community get assigned to the community

* Re-add remove attachment to themes controller; remove active from available attributes on theme

* Disallow Super Admins from viewing profile pags of Place, Speaker, Story, and Curriculum Story

* Create before action so show actions are actually authorized

* ensure Stories, Speakers, and Places all have import form

* Disallow (view and controller) users from editing their own role unless admin/super admin

* Super Admin can create first admin user in Community when created

* Remove erroneous byebug

* restrict, for now, super admins from accessing /home

* Brand devise login/session#new page for a better Terrastories landing experience

* Add concept of viewer role which can belong to a community but only see public / anon stories