Completed test for IntroPopup (#595)
Added a test for IntroPopup that uses Enzyme for element rendering.

One test checks that the shallow render matches a stored snapshot,
which I think is the most efficient way to check its basic rendering,
as this will ensure it has the right elements, structure, and content
and is easily reviewed and updated if there are future design changes.

There are also checks that the component uses the supplied props as
expected, which currently is no props used and sets the isPopped state
to true.  Added tests for passing isPopped prop, for a possible future
implementation of time-delay popup or programmed popup, which are currently
skipped so they won't cause any failures.

Also included is a mock of I18n translation and a test that the rendering
is done via I18n.  The mock includes a mock implementation to assist
with shallow rendering inspection that identifies what content would
be rendered for a certain translation.