Update gems (#643)
* Update gems

* Symbolize route builder

A vulnerabilty in polymorphic paths in Rails that was patched now always
warns when a string input is passed in to route building.

For example:

In Administrate, we use `page.resource_name` or `page.resource_path` to
generate paths. Unfortunately, these both return strings which is
flagging error.

The purpose is to block possible nefarious user input generating URLs
that are problematic. In our case, these are administrate managed
resources and we can safely cast to symbol.