Ensure Dashboard views are ready for translation
Not all keys are translated yet. I did my best to copy them over from
existing translations, but there are English translations in certain
places due to not having a one-to-one existing translation.

Note: Some translations are duplicated between Administrate and
Dashboard; if you are ADDING: please add to DASHBOARD. This is what will
be used long term; the administrate ones will be removed when
administrate is removed.

Model validations were updated and translated as well, which is not
something that was the case before.