Remove unused gem: dotenv-rails
While we do have a `.env` file, it's used for Docker purposes and not at
the root of the Rails application. Since we are not loading .env into
the app in this way, we can remove the gem.

Plus, with the use of Docker, we should be using .dockerenv or
specifying env in the Dockerfile or Compose files.