General cleanup and upgrades (#867)
* Remove unused gem: acts-as-taggable

It appears the code that used this gem was moved a long time ago, but
the gem lingered.

* Remove unused gem: dotenv-rails

While we do have a `.env` file, it's used for Docker purposes and not at
the root of the Rails application. Since we are not loading .env into
the app in this way, we can remove the gem.

Plus, with the use of Docker, we should be using .dockerenv or
specifying env in the Dockerfile or Compose files.

* Remove font-awesome-rails gem

This appears to be have been added, but never made fully functional. A
quick search for the font awesome `fa_*` helpers and finding none,
removing this gem seemed fine.

If we ever wish to utilize font awesome in the future, we can add it
then.

* Remove rails-controller-testing gem

* Reorganize and document gems in Gemfile

* Remove obsolete custom DeviseFailureApp.

This is left over from when we attempted to run Devise on a subdomain.
We abandoned this shortly after but the code lingered. Removing for now.