feat: Add automatic onboarding for offline commmunity
This adds an automatic onboarding detection if a community is booting up
Terrastories for the first time. It's determined by if the app is in
"offline" mode (either enabled by using an offline/local tileserver or
by running with RAILS_ENV=offline) AND the database does not detect an
existing Community.

It also, at the moment, adds a just-in-case Super Admin account.