Moves Sponsor Logos and Background Image to Community
In an effort to make "theme" map settings and put community settings
directly on the community, this moves the sponsor logos and the
background image directly onto the Community.

This is done via a data-only migration which updates the active storage
attachments to switch from Theme to Community (and back on a rollback).

For backwards compatibility:

1. The Theme page for Admin's now accepts nested attributes for the
   background image and sponsor logos. This will ensure current behavior
   appears to stay the same, even though the backend storage is changed.
2. If a Community has `split_settings` feature enabled, the welcome area
   settings in Theme is hidden and is instead available in the Settings
   tab in the sidebar.

This also fixes an issue where "Settings" was not showing up since it
was checking if Split Settings was enabled for ALL and not just the
current community.

And lastly, this removes the `set_theme` before actions that were
required in order to display the splash welcome screen background and
logos, since now we just look it up from the Community record.