Feat: Community Settings (#888)
* Add slug to Community

This slug will be used in URL's for public interface to avoid exposting
ID's or other database information.

* Reverse Theme <> Community relationship

Before, Community belonged to a theme. This caused some weird behavior
on how the Theme was created when a Community was created (Theme had to
be created/built first). Now, a Theme belongs to a Community and is
created after a Community is created.

* Remove theme_id from community

* Moves Sponsor Logos and Background Image to Community

In an effort to make "theme" map settings and put community settings
directly on the community, this moves the sponsor logos and the
background image directly onto the Community.

This is done via a data-only migration which updates the active storage
attachments to switch from Theme to Community (and back on a rollback).

For backwards compatibility:

1. The Theme page for Admin's now accepts nested attributes for the
   background image and sponsor logos. This will ensure current behavior
   appears to stay the same, even though the backend storage is changed.
2. If a Community has `split_settings` feature enabled, the welcome area
   settings in Theme is hidden and is instead available in the Settings
   tab in the sidebar.

This also fixes an issue where "Settings" was not showing up since it
was checking if Split Settings was enabled for ALL and not just the
current community.

And lastly, this removes the `set_theme` before actions that were
required in order to display the splash welcome screen background and
logos, since now we just look it up from the Community record.

* Fix: reset.scss overwrote font-style on html tags incorrectly.

* Replace `associated_updated_at` w/ touch: true

Rather than building associated_updated_at by doing several trips to the
database to look up most recent updated_at date; we can utilize Rails
belong_to relation option `touch: true` to update community every time a
change is made to a nested model that we want to track on the Community.

* include community sorts in filter form for Super Admin

* remove theme from community seed

* Fix: allow split settings community update to save background img and sponsor logos

* Fix: move delete attachments from Theme controller to Community controller

* Fix: fk and orphaned theme migration

* Demo: attach example bg image in seeds

* Demo: attach example sponsor logo in seeds

* Fix: syntax error in Welcome#show

Co-authored-by: Rudo Kemper <31662219+rudokemper@users.noreply.github.com>