Switch from using i18n-js to react-i18next
The first of a few refactors to stop loading the front-end map portion
("Home") using BOTH webpacker AND asset pipeline / sprockets. The split
is confusing and has caused some issues with where to include JS /
packages / etc.

i18n-js utilize Rails translations and allowed them to be converted and
loaded into JS in a rails-y way. v3 of the gem, this had a very
"magical" feel and did not require a separate package, though one was
optionally available. v4 of them required both the gem AND the package,
plus an additional build-time CLI to export the translation files for
use in a production environment.

The v3 to v4 ugprade would require a lot of additional work and
maintenance overhead and our long-term goal is to move translations into
an external / no-code oss platform for shared use. As such, this PR
separates out the React portion of translations from the Rails
translations and defines them to use in JSON files for react i18next.