Update Devise
Devise 4.9 is Ruby 3 ready. Since we are still using ujs (and not
turbo), most of the changes are irrelevant to our setup.

For ease in future updates, this makes sure that links relying on UJS,
both in and outside of devise, are both ujs and turbo compatible. This
also opts in to the new Devise configuration.