Upgrades Ruby, Node, and dependencies (#908)
* Upgrade webpacker to latest

* Switch from using i18n-js to react-i18next

The first of a few refactors to stop loading the front-end map portion
("Home") using BOTH webpacker AND asset pipeline / sprockets. The split
is confusing and has caused some issues with where to include JS /
packages / etc.

i18n-js utilize Rails translations and allowed them to be converted and
loaded into JS in a rails-y way. v3 of the gem, this had a very
"magical" feel and did not require a separate package, though one was
optionally available. v4 of them required both the gem AND the package,
plus an additional build-time CLI to export the translation files for
use in a production environment.

The v3 to v4 ugprade would require a lot of additional work and
maintenance overhead and our long-term goal is to move translations into
an external / no-code oss platform for shared use. As such, this PR
separates out the React portion of translations from the Rails
translations and defines them to use in JSON files for react i18next.

* Remove plyr-rails gem

This appears to be unused despite being loaded. Removing the gem.

If we decide we want to use / link up Plyr in the future, we can do so
using the NPM package for the map sidebar; or re-add it for the
dashbaord (where it is not current loaded).

* Use npm mapbox-gl for React map

This does NOT remove the mapbox-gl gem as that is used by the dashboard;
however, it is removed from the application styles and javascript
includes and is instead included in package.json and imported for usage.

This also required converting the minimap to an es6 module.

* Switch i18n setup to its own file

* Bundle pause_all_videos in application pack

* Bundle React/Map styles in Webpack; not sprockets

* Lock mapbox-gl version to the same as Rails gemfile version (2.9.0)

* Add comments and remove extraneous consoles.

* Fix deprecration error: DEPRECATION WARNING: Rendering actions with '.' in the name is deprecated

* Upgrade Ruby 2.5.9 -> 2.6.0

* Upgrade Ruby 2.6.0 -> 2.6.10

* Update bundler

* Config change

* Set versions for gems

* Only require rails modules that we use

* bundle update rails

* bundle update pg

* fixup: moduleize

* Switch from CSS in Webpacker to CSS Bundling

(my bad on moving everything to webpacker...only to learn after that it
was a complete waste of time...)

* require relative

* Upgrade Ruby 2.6.10 -> 2.7.0

* Ruby 2.7.0 -> 2.7.7 (latest patch in 2.7 before 3.0)

* Update React-Select

* Update aws-sdk-s3

* Update Devise

Devise 4.9 is Ruby 3 ready. Since we are still using ujs (and not
turbo), most of the changes are irrelevant to our setup.

For ease in future updates, this makes sure that links relying on UJS,
both in and outside of devise, are both ujs and turbo compatible. This
also opts in to the new Devise configuration.

* Update devise-i18n to match with devise updates

* Update Flipper; also fix missing community -> current_community

* Update geocoder

* Update Listen

* Update Pundit

* Update comments

* Upgrade @babel/core

* Bump i18next

* Only include one source for Mapbox

Before we bundled both Mapbox Rails (gem) and Mapbox gl (npm). This
removes the Rails version (which was a wrapper for the js file, and
frequently behind).

Asset files correctly add mapbox-gl to webpacked version (import) and
sprokets (//= require mapbox-gl).

* Upgrade mapbox-gl

* Add webpack-dev-server to dev bootup

* Remove tileserver fonts

* Add tileserver specific gitignore

* rails credentials key is optional, even in production

* Remove ActionCable

* Remove root level .dockerignore (unused)

* Remove unused start script

* Readd mailer, required for devise

* Improve offline configuration and add comments

* Use single Dockerfile with mulit-stage targets for dev

* Remove now obsolete standalone dockerfile

* Add workflow to build and push devcore image to Docker

---------

Co-authored-by: Laura Mosher <lauramosher@users.noreply.github.com>