Offline Mode: Generate secret_key_base
Between the creation of Rails.env.offline, removal of Rails.credentials
and the requirement of RAILS_MASTER_KEY, and updates, RAILS_ENV=offline
stopped booting due to missing SECRET_KEY_BASE.

Rather than requiring folks setting up the full development mode to have
one created automatically, this updates the environments/offline.rb file
to generate a new one and store it in a tmp file accessible by the app
(and stored in Docker, similar to development). This also removes the
need to generate and store one in the Docker image (which would be used
by all communities using that image), or adding even more steps to
generate the string and add it.

It mirrors the implementation Rails uses for Rails.env.test and
Rails.env.development.

It also accounts for any setups that DO utilize an ENV variable. So if
one is set, it uses it, otherwise it creates a new one.