Fix Offline Mode Configuration & Setup (#912)
* Use compose host (db) in default offline DATABASE_URL configuration.

* Offline Mode: Generate secret_key_base

Between the creation of Rails.env.offline, removal of Rails.credentials
and the requirement of RAILS_MASTER_KEY, and updates, RAILS_ENV=offline
stopped booting due to missing SECRET_KEY_BASE.

Rather than requiring folks setting up the full development mode to have
one created automatically, this updates the environments/offline.rb file
to generate a new one and store it in a tmp file accessible by the app
(and stored in Docker, similar to development). This also removes the
need to generate and store one in the Docker image (which would be used
by all communities using that image), or adding even more steps to
generate the string and add it.

It mirrors the implementation Rails uses for Rails.env.test and
Rails.env.development.

It also accounts for any setups that DO utilize an ENV variable. So if
one is set, it uses it, otherwise it creates a new one.

* Lock Dockerfile image tag to minor ruby patch

* Pre-compile assets for offline mode setup

This updates bin/setup to precompile assets of offline mode setup is
selected. It's prefixed by offline-assets so as to not interfere with
other precompilation tasks.

Developers using RAILS_ENV=offline will need to beware that changes to
assets will not reflect without re-compiling between changes. It's
recommended to test offline tileserver maps in development mode using
OFFLINE_MAP_STYLE or USE_LOCAL_MAP_SERVER for testing with
RAILS_ENV=development instead.

---------

Co-authored-by: Laura Mosher <lauramosher@users.noreply.github.com>