Implement location picker map for places
https://github.com/Terrastories/terrastories/issues/805

This is a very similar implementation as the one for [themes](https://github.com/Terrastories/terrastories/blob/bb0784bd162dc2095b75afb7d01f0841ee0dadb4/rails/app/views/dashboard/themes/edit.html.erb#L163) (using the same 2 column CSS and set of event listeners), and includes functionality like https://github.com/Terrastories/terrastories/pull/921 as well.

I verified this works offline as well, except we need to merge https://github.com/Terrastories/terrastories/pull/920 to fix the CSS issue.

I also verified it works for both Places #new and #edit (they share the _form partial)

As the issue mentioned, I locked the marker to the center of the map
instead of making it independently draggable. One of the consequences of
this is that it's a bit weird to set a point to one of the latitude
extremes (i.e. -90 or 90). Even editing this manually will get the value
reset to the center of the current map viewport depending on the zoom
level.