Move the map center after editing the lat/long fields (#921)
When dragging the map around, the center coordinates get synced to the
input fields, but not the other way around. This PR implements syncing
in the reverse direction in case the user wants to edit the coordinates
manually but wants to see where that is on the map visually.

I originally tried to listen for [input event](https://developer.mozilla.org/en-US/docs/Web/API/HTMLElement/input_event), but it would fire as the user is typing and modify the value to include the 5 fixed decimal points, which would effectively prevent the user from entering any values more than a single digit. So I am listening for [change event](https://developer.mozilla.org/en-US/docs/Web/API/HTMLElement/change_event) instead, which has the small negative that it won't fire until the input field loses focus.

Also there is a small refactor to reuse the variables for the input fields

I noticed this while working on https://github.com/Terrastories/terrastories/issues/805