Fix zoom for edit place
It looks like theme is created for a community in an after_create, but
it's possible for center_lat, center_long, or zoom to be null, so add
default fallbacks. I would prefer to make these columns on theme not
null with a default, but not sure what the implications of that might
be.