Adds additional check for public community settings
On the off chance that a community has the public community feature
enabled but does not yet have a slug assigned, this hides the whole box
as if the feature is not enabled.

This ensures that there isn't an awkward save loop where everytime you
check the box and click submit it goes back to unchecked.

At the time of this change, only one community has no slug assigned.

Once Public Communities is no longer an opt-in feature, we'll ensure all
communities have a slug and remove both checks.