fix: only attempt to generate static map when access token is available
This is mostly as a JustInCase measure; and especially for possible
issues with local development.

Before we only checked that a community had opted into a public
interface. Now we also check that they have an access token set (either
community level or default) through the theme method. And then in the
event that the token is bad or expired or some other network issue
occurred, we rescue any standard error when trying to IO stream the file
and attach it and continue as if nothing bad happened.

The static map is to ensure that our default provided mapbox API token
is not accidentally innundated with requests on the loading screen of
Explore where we might hit API rate limits.