Feature: Public Communities (#934)
* Add display image to community

* Add display image and bg img/sponsor logos to community key in models locales

* Rename subdir application controllers to base controller

* Feat: Public Communities

This adds the API endpoints required for public communities. The ability
to explore the maps of communities that mark themselves as public is a
separate app from this one.

* Update CORS to only allow /api access to specific configured origins

* Ensure both limits for resize_to_limit are set

* Remove server renderings pack; not used

* Update filters

Before, filters would return options for ALL stories, regardless of
story permission level.

Now, filters are, by default, built using only publicly available
stories. Optionally, an array of permission levels can be passed as the
first argument of the method to expand those options.

* Make the default URL stuff configurable

* Use representation_url for variants

* Remove port from default URL options for routing

* Switch API points from a FeatureCollection to an array of PointFeatures.

* Add Thumbnail photo to Place GeoJSON

- Added thumbnail flag to photo_url.
- Current size applied to thumbnail version.
- Added full-size-ish size for standard photo size.

* Only send filterable category keys to front end api request

* fix: photo and place name audio urls for api access

* feat(public communities): generate static image from mapbox

Public communities home page list benefits from displaying a snapshot of
the community map when community opts in to public display. This will
generate a static map when a community is saved and public is enabled
AND when a community theme is updated.

* fix: restricted stories showing on public places

* feat: send story media content types for easy type display

* fix: speaker communities filters were duplicated

* refactor: move tooltips to their own reusable set of styles

* feat: override slug setter to ensure unique slugs

* feat: allow super admins to set and update slugs for communities

* feat: bubble slug and public on super admin community show page

* fixup: override setter

* Adds additional check for public community settings

On the off chance that a community has the public community feature
enabled but does not yet have a slug assigned, this hides the whole box
as if the feature is not enabled.

This ensures that there isn't an awkward save loop where everytime you
check the box and click submit it goes back to unchecked.

At the time of this change, only one community has no slug assigned.

Once Public Communities is no longer an opt-in feature, we'll ensure all
communities have a slug and remove both checks.

* feat: show and copy to clipboard public url

When the feature is active for a community and the community has opted
into public consumption, a box will appear that shows the public URL
(currently hard-coded for PRODUCTION; so this will not work for staging
at this time).

* fix: only attempt to generate static map when access token is available

This is mostly as a JustInCase measure; and especially for possible
issues with local development.

Before we only checked that a community had opted into a public
interface. Now we also check that they have an access token set (either
community level or default) through the theme method. And then in the
event that the token is bad or expired or some other network issue
occurred, we rescue any standard error when trying to IO stream the file
and attach it and continue as if nothing bad happened.

The static map is to ensure that our default provided mapbox API token
is not accidentally innundated with requests on the loading screen of
Explore where we might hit API rate limits.

* fix: api spec descriptors

* enhancement: use AR's bool attr? for clarity

* chore: update Community factory with newer changes to slug

* Add rescue to Themes controller as well

* chore: extract eye svg icon into shared icon symbols

---------

Co-authored-by: Laura Mosher <lauramosher@users.noreply.github.com>