fix(API): Catastrophic 500 when media cannot be processed (#976)
* Red Tests

* fix(API): Explore shows infinite loading

When an otherwise representable media attachment is processed for a
preview or image and errors, the request fails catastrophically
resulting in the Explore interface looking like it's loading forever.

It was determined that this was due to the underlying processing failure
and, as such, the check for `.representable?` was insufficient to catch
errors.

For now, we simply added a test that raises the error manually for
invariable images. See https://api.rubyonrails.org/classes/ActiveStorage/Blob/Representable.html#method-i-variant

Our code currently catches any ActiveStorage::Error to reduce friction
points since the Explore code has guardrails in place to display other
media when a preview thumbnail is unaavilable.

---------

Co-authored-by: Laura Mosher <lauramosher@users.noreply.github.com>