Utilize per-community Mapbox settings when set
This utilizes the new default Map configuration so we can better test
the offline/online and map configuration.

When offline, MapLibre and default map package is used, regardless of
community settings.
When online, if a community has Mapbox configured, even if MapLibre is
the default map renderer, that community's map will render with Mapbox